requirements.txt
Flask==3.0.3
gunicorn==23.0.0
numpy==1.24.4
python-dotenv==1.0.1

